+++ changes.md
@@ -1,3 +1,4 @@
 
 - ConditionalEffects disabled
 - FogOfWar is not functional
+- EchoExtender configs work by name (White/Yellow/Red)